aapits, osunixxf: Semaphore emulation layer improvements.
Semaphore OS calls are use by lock and mutex OS calls. After lock and mutex
creation it can be Acquire without any early signaling.
Semaphore need to be signaled to be acquired by the nature of semaphore.
So this commit enable for using semaphore by lock, mutex and directly as well.